feat:migrated to typescript and added a ci.yml file
