--- README.md
@@ -1,2 +1,0 @@
-Hello
-This is a a very simple website about a shopping website
