+++ package.json
@@ -3,15 +3,17 @@
   "version": "1.0.0",
   "main": "server.js",
   "scripts": {
-    "start": "nodemon server.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
+  "start": "ts-node-dev --respawn --transpile-only server.js",
+  "typecheck": "tsc --noEmit",
+  "build": "tsc",
+  "test": "echo \"Error: no test specified\" && exit 1"
+},
   "author": "Bello Ayoola",
   "license": "ISC",
   "description": "",
   "dependencies": {
     "axios": "^1.9.0",
-    "bcrypt": "^5.1.1",
+    "bcrypt": "^6.0.0",
     "cloudinary": "^2.5.1",
     "cookie-parser": "^1.4.7",
     "cors": "^2.8.5",
@@ -24,10 +26,23 @@
     "mongoose": "^8.9.2",
     "morgan": "^1.10.0",
     "multer": "^1.4.5-lts.1",
-    "node-cron": "^3.0.3",
-    "nodemailer": "^7.0.2",
+    "node-cron": "^4.6.0",
+    "nodemailer": "^9.0.3",
     "redis": "^4.7.0",
     "socket.io": "^4.8.1",
     "validator": "^13.12.0"
+  },
+  "devDependencies": {
+    "@types/bcrypt": "^6.0.0",
+    "@types/cookie-parser": "^1.4.10",
+    "@types/cors": "^2.8.19",
+    "@types/express": "^5.0.6",
+    "@types/jsonwebtoken": "^9.0.10",
+    "@types/morgan": "^1.9.10",
+    "@types/multer": "^2.2.0",
+    "@types/node": "^26.1.1",
+    "@types/node-cron": "^3.0.11",
+    "ts-node-dev": "^2.0.0",
+    "typescript": "^5.9.3"
   }
 }
