+++ tsconfig.json
@@ -1,0 +1,25 @@
+{
+    "compilerOptions": {
+    /* Let TS and JS live side by side during the migration */
+    "allowJs": true,
+    "checkJs": false,
+
+    /* Output */
+    "target": "ES2020",
+    "module": "CommonJS",
+    "outDir": "dist",
+    "rootDir": ".",
+
+    "esModuleInterop": true,
+    "resolveJsonModule": true,
+    "skipLibCheck": true,
+    "forceConsistentCasingInFileNames": true,
+
+    "strict": true,
+    "noImplicitAny": true,
+
+    "sourceMap": true
+  },
+    "include": ["**/*.ts", "**/*.js"],
+    "exclude": ["node_modules", "dist", "uploads"]
+}
